+++ docs/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Watchdog timer (WDT)
 - Ambient starts `Alarm_ambient_dimming_duration` before the alarm triggers,
   so that it already has full brightness when the alarm triggers.
+- Countdown timer feature.
 
 ### Changed
 - `dis` command in CLI changed to `en-off` for consistency